requirements.txt
gunicorn==20.0.4
uvloop==0.15.2
python-jose==3.2.0
passlib==1.7.4
pydantic==1.8.1
python-multipart==0.0.5
bcrypt==3.2.0
numpy==1.19.2
#.....................# 
google-auth==1.33.0
google-cloud-storage==1.41.0
Keras-Preprocessing==1.1.2
tensorflow==2.4.0
keras==2.4.0
scikit-image==0.18.2 
scikit-learn==0.24.2 
matplotlib==3.4.2

